sips/requirements.txt
Django==1.10.4
django-bootstrap3==7.1.0
django-datetime-widget==0.9.3
lxml==3.7.1
olefile==0.43
Pillow==4.0.0
psycopg2==2.6.1
python-docx==0.8.6
pytz==2016.10
reportlab==3.3.0
